ui: move libraries to lib/
It doesn't make sense to keep non-component libraries under
src/components, so move them to a new lib/ directory.
