+++ src/js/frontend/src/lib/d3Chart.js
